// web/src/lib/generated/hashpower-calculator/HashpowerCalculatorResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CacheMode } from "./CacheMode";
import type { CalculatorInputs } from "./CalculatorInputs";
import type { CalculatorResults } from "./CalculatorResults";
import type { CalculatorWarning } from "./CalculatorWarning";
import type { MarketSnapshot } from "./MarketSnapshot";

export type HashpowerCalculatorResponse = { inputs: CalculatorInputs, market: MarketSnapshot, results: CalculatorResults, warnings: Array<CalculatorWarning>, stale: boolean, cacheMode: CacheMode, };